Accept list of tomls as an input for TestParser
This will allow making verify modes smarter by gathering all available
Test TOMLs and passing them as input when Scenario is verified.
