+++ src/cloudai/cli/handlers.py
@@ -231,9 +231,7 @@ def handle_verify_tests(args: argparse.Namespace) -> int:
     for test_toml in test_tomls:
         logging.info(f"Verifying {test_toml}...")
         try:
-            parser = TestParser(Path(), None)  # type: ignore
-            parser.current_file = test_toml
-            parser.load_test_definition(toml.load(test_toml))
+            Parser.parse_tests([test_toml], None)  # type: ignore
         except Exception:
             nfailed += 1
