+++ tests/test_test_definitions.py
@@ -18,7 +18,7 @@ from pathlib import Path
 
 import pytest
 import toml
-from cloudai import Registry, TestDefinition, TestParser
+from cloudai import Parser, Registry, TestDefinition
 from cloudai.test_definitions import ChakraReplayCmdArgs, NCCLCmdArgs, NCCLTestDefinition
 
 TOML_FILES = list(Path("conf").glob("**/*.toml"))
@@ -64,9 +64,7 @@ def test_all_tests(toml_file: Path):
     template_name = toml_dict["test_template_name"]
     assert template_name in registry.test_definitions_map, f"Unknown test template: {template_name}"
 
-    parser = TestParser(Path(), None)  # type: ignore
-    parser.current_file = toml_file
-    parser.load_test_definition(toml.load(toml_file))
+    Parser.parse_tests([toml_file], None)  # type: ignore
 
 
 def test_chakra_docker_image_is_required():
